+++ pkg/commands/service/backend/create.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"net"
 
-	"github.com/fastly/go-fastly/v14/fastly"
+	"github.com/fastly/go-fastly/v15/fastly"
 
 	"4d63.com/optional"